// src/generated/runtimeDescriptions.ts
export interface RuntimeParameter {
	name: string;
	type?: string;
	description?: string;
	default?: string;
}

export interface RuntimeMember {
	name: string;
	kind: 'property' | 'method';
	signature?: string;
	description: string;
	fullDescription?: string;
	returnType?: string;
	parameters?: RuntimeParameter[];
}
// ...
export const flowMembers: Record<string, RuntimeMember> = {
	"function_dir": {
		name: "function_dir",
		kind: "property",
		description: ``,
	},
	"current_step": {
		name: "current_step",
		kind: "property",
		description: `The name of the step we're currently in`,
		returnType: `str`,
	},
	"functions": {
		name: "functions",
		kind: "property",
		description: `The functions available to the flow`,
		returnType: `FlowFunctionExecutor`,
	},
	"goto_step": {
		name: "goto_step",
		kind: "method",
		signature: `(step_name: str, label: str | None = None)`,
		description: `Trigger a transition to a different step`,
		fullDescription: `Trigger a transition to a different step

Args:
step_name: name of the step to transition to
label: used for rendering the condition where relevant but no runtime effect`,
		parameters: [
			{ name: "step_name", type: `str`, description: `name of the step to transition to` },
			{ name: "label", type: `str | None`, description: `used for rendering the condition where relevant but no runtime effect`, default: "None" },
		],
	},
};